chore: change default endpoint to grpc
