+++ src/jenesis/cmd/add/profile.py
@@ -3,7 +3,7 @@ import os
 import toml
 
 from jenesis.config import ConfigurationError
-from jenesis.contracts.networks import fetchai_testnet_config, fetchai_localnode_config
+from jenesis.network import fetchai_testnet_config, fetchai_localnode_config
 
 def run(args: argparse.Namespace):
